Add BaseTestClass & set as a dependency.
From the recent upgrade to php 7 I can see that every file had to be changed to reflect the
phpunit version. By extending a class we can put that change in one place.

I'm actually working on php 5.6 so this allows me to run the tests on that
version with a low level of tinkering