Add Factory methods.
Per the unit tests this adds the following factory pattern

This is based on the way the Omipay extension does it. You might not want
to includete swiftmailer in the composer.json / in the factory test, but
I wanted to demonstrate how this could work with packages that leveraged
your packages, without them actually having to be in your package. Once
again that is based on how Omnipay does it.