Add mass mailing functions.
I have been working on adding support for the retrieval of details about mass mailings and the actions taken by recipients of them. The principles I have been working off are that the calling code should need to know very little about the mailing provider code, and that the code should support get-on-demand, by which I mean that some details about mailings may require a different api call to retrieve them, and this should be retrieved when required by the Omnimail integration, without the calling code needing to know. I have based my approach on Omnipay, who I believe have done a good job doing this for payment processors.

Per the Omnipay approach I have given each action a Request and a Response object. The logic here is that the Mailer method returns a Request Object. Various standardised properties can be set on the RequestObject and then when getResponse() is called it returns a response object. The response object represents multiple mailings, or recipients, and is an object that extends  BaseResponse, which in turn extends the array object. Using an array object allows it to be iterated with a foreach, but also allows properties and methods directly on the response object.

I have put code examples in the Readme. My extension is here https://github.com/eileenmcnaughton/omnimail-silverpop

Some things I am still pondering
1) one of the weaknesses of Omnipay (which I have taken much inspiration from) is the fact the calling code needs to know the names of the credentials fields. If
the calling code is implementing multiple providers this can be a pain. I think maybe a setCredentials function might make sense, where incoming credentials
would simply be numerically-indexed. The calling app still needs to store the credentials in order, but not know if they are apiKey or signature or userName etc.

2) In the case of silverpop the getRecipients action is 2 part - it requests a download and then retrieves it. I added an isCompleted
function so the code can loop around until it is ready. In this case if you wish to not have the php code sleeping & do it in 2 separate
processes there are some parameters that are available & need to be used in the next process. I have used getRetrievalParameters to get these
and setRetrievalParamters to set them on the next instance.